LPS-66316 Conflicting package mail.util: move Attachment
