+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/attachment/AttachmentHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mail.util;
+package com.liferay.mail.attachment;
 
 import java.io.InputStream;
